stflow Rename and reorganise tests
